+++ README.md
@@ -1,1 +1,13 @@
 # parts-undefined
+
+## Docker
+
+Build the Dockerfile
+```sh
+docker build -t parts-unknown .
+```
+
+Run the container
+```sh
+docker run -it --rm -p 3000:3000 --name parts-unknown parts-unknown
+```
