+++ back-end/app/controllers/api/blog_comments_controller.rb
@@ -1,2 +1,4 @@
+module Api
 class BlogCommentsController < ApplicationController
+end
 end
