--- back-end/app/controllers/application_controller.rb
@@ -1,29 +1,0 @@
-class ApplicationController < ActionController::API
-  # skip_before_action :verify_authenticity_token
-    include ActionController::Cookies
-      
-    rescue_from ActiveRecord::RecordInvalid, with: :render_unprocessable_entity_response
-    rescue_from ActiveRecord::RecordNotFound, with: :render_not_found_response
-
-    # before_action :authorize
-
-    def hello_world
-      session[:count] = (session[:count] || 0) + 1
-      render json: { count: session[:count] }
-    end
-
-    def authorize
-      @current_user = User.find_by(id: session[:user_id])
-  
-      render json: { errors: ["Not authorized"] }, status: :unauthorized unless @current_user
-    end
-    
-    def render_unprocessable_entity_response(invalid)
-      render json: { errors: invalid.record.errors.full_messages }, status: :unprocessable_entity
-    end
-    
-    def render_not_found_response
-      render json: {error: "Not found"}, status: :not_found
-    end
-    
-end
