+++ client/src/components/ForumCountry.js
@@ -12,7 +12,7 @@ export default function ForumCountry({ countries, setCountries, setSearch }) {
   // const [countriesSort, setCountriesSort] = useState([]);
 
   // const { data: countries, isLoading } = useQuery(["country"], () => {
-  //   return Axios.get("http://localhost:3000/countries").then((res) => res.data);
+  //   return Axios.get("/api/countries").then((res) => res.data);
   // });
 
   const lastPostIndex = currentPage * postsPerPage;
