+++ client/src/components/LandingPage.js
@@ -11,7 +11,7 @@ import Axios from "axios";
 
 export default function LandingPage() {
   const { data: countries, isLoading } = useQuery(["country"], () => {
-    return Axios.get("http://localhost:3000/countries").then((res) => res.data);
+    return Axios.get("/api/countries").then((res) => res.data);
   });
 
   let slides = [
