+++ client/src/components/Login.js
@@ -7,7 +7,7 @@ export default function Login({ setUser }) {
 
   function handleSubmit(e) {
     e.preventDefault();
-    fetch("/login", {
+    fetch("/api/login", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
