+++ client/src/components/WhereToNext.js
@@ -9,7 +9,7 @@ export default function WhereToNext() {
   const navigate = useNavigate();
 
   const { data: destinations, isLoading } = useQuery(["destination"], () => {
-    return Axios.get(`http://localhost:3000/randomdestinations`).then(
+    return Axios.get(`/api/randomdestinations`).then(
       (res) => res.data
     );
   });
