+++ client/src/components/quizzes/quizTwo/Result.js
@@ -8,7 +8,7 @@ function Result(props) {
   const navigate = useNavigate();
 
   const { data: destinations, isLoading } = useQuery(["destination"], () => {
-    return Axios.get("http://localhost:3000/destinations").then(
+    return Axios.get("/api/destinations").then(
       (res) => res.data
     );
   });
